backend/requirements.txt
fastapi[standard]
tinydb
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
